% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ALA_data.R
\name{aggregate_ALA_data}
\alias{aggregate_ALA_data}
\title{Aggregate ALA data to grid cells}
\usage{
aggregate_ALA_data(ALA.filtered.data, domain.mask, agg.radius.ncells = NULL,
  output.folder = NULL, output.name = "aggregated_taxa_data",
  verbose = TRUE)
}
\arguments{
\item{ALA.filtered.data}{(dataframe) A dataframe of appropriate format held in R working environment.}

\item{domain.mask}{(raster layer) A raster layer specifying the analysis domain}

\item{agg.radius.ncells}{(float) A spatial radius, specified as number of grid cells, over which to aggregate records. If NULL, no radial aggregation is performed. Default: NULL.}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string), A name to use in saving the outputs. Default: 'aggregated_taxa_data'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\value{
Dataframe. Also writes this dataframe to file if output.folder is specified.
}
\description{
Take filtered occurrence data and group their position to grid cell centres.
}
\examples{
output = aggregate_data(my.data, my.raster)

}
