% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dissimilarities.R
\name{calculate_dissimilarities}
\alias{calculate_dissimilarities}
\title{Calculate dissimilarities}
\usage{
calculate_dissimilarities(pairs.table = Pairs.Table.Dens,
  composition.data = Selected.records, output.folder = NULL,
  output.name = "pairs_table_dissim", verbose = TRUE)
}
\arguments{
\item{pairs.table}{(dataframe) A dataframe holding the site-pairs. This is the first six columns of a gdm input table.}

\item{composition.data}{(dataframe) A dataframe holding the final species composition data: the records of all species in all grid cells to be used for modelling.}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. Default: 'pairs_table_dissim'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\description{
Calculates the compositional dissimilarity between all specified site-pairs. Fills the dissimilarities in the input dataframe 'pairs.table'. Note this uses an approach that trades processing time for memory efficiency, so it can handle very large numbers of site pairs on any machine.
}
\examples{
output = calculate_dissimilarities(My.pairs.table, My.composition.data, output.folder = 'C:/Users/processed_data', output.name = 'My.sitepair.data')

}
