% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gdm_deviance.R
\name{calculate_gdm_deviance}
\alias{calculate_gdm_deviance}
\title{Calculate GDM deviance for observed & predicted dissimilarities}
\usage{
calculate_gdm_deviance(pU, pY)
}
\arguments{
\item{pU}{(float) A vector of predicted dissimilarity values, of same length as pY.}

\item{pY}{(float) A vector of observeded dissimilarity values, of same length as pU.}
}
\value{
A single value (float) being the deviance.
}
\description{
Calculate GDM deviance for observed & predicted dissimilarities. Can be used for assessing cross-validation data. Translated from the c++ function 'CalcGDMDevianceDouble()' in the file 'NNLS_Double.cpp' from the GDM R package.
}
\examples{
output = calculate_gdm_deviance(My.pU, My.pY)

}
