% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_filepath}
\alias{check_filepath}
\title{Ensures filepath has trailing forwardslash}
\usage{
check_filepath(filepath)
}
\arguments{
\item{filepath}{(string)}
}
\value{
string (filepath with trailing forwardslash)
}
\description{
Adds trailing forwardslash to filepath if not present so that the filepath can be added to for writing.
}
\examples{
fn = checkfilepath('C:/Users')
}
