% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{counter}
\alias{counter}
\title{Simple counter to monitor a loop}
\usage{
counter(msg = "Doing loop ", vec = NULL, iterable = NULL)
}
\arguments{
\item{msg}{(string) Default: 'Doing loop.' Whatever needs to be printed...}

\item{vec}{(vector) A vector of things to iterate through with each i in vec being printed as progress is made. Default: NULL.}

\item{iterable}{An object that is iterated over in a loop (i.e. the i in the above example)}
}
\value{
Output to console
}
\description{
Counter that doesn't clog the console but enables progress to be monitored
}
\examples{
for (i in LETTERS[1:26]) counter(iterable = i)
}
