% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_env_data.R
\name{extract_env_data}
\alias{extract_env_data}
\title{Extract environmental data}
\usage{
extract_env_data(ALA.composition.data, environment.stk, output.folder = NULL,
  output.name = "site_env_data", verbose = TRUE)
}
\arguments{
\item{ALA.composition.data}{(dataframe) A dataframe holding the final species composition data.}

\item{environment.stk}{(raster stack) A stack of rasters from which environment data will be extracted.}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string), A name to use in saving the outputs. Default: 'filtered_data'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\value{
Dataframe.
}
\description{
Extracts environmental data for specified locations, from a set of specified spatial grids. Returns a dataframe (site x env).
}
\examples{
output = extract_env_data(My.composition.data, My.env.stk, output.folder = 'C:/Users/processed_data', output.name = My.filtered.ALA.data, domain.mask = Aust.ras)

}
