% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ALA_data.R
\name{filter_ALA_data}
\alias{filter_ALA_data}
\title{Filter downloaded ALA data}
\usage{
filter_ALA_data(ALA.download.data, output.folder = NULL,
  output.name = "filtered_data", domain.mask = NULL, earliest.year = 1970,
  spatial.uncertainty.m = 2000, select.fields = NULL, verbose = TRUE)
}
\arguments{
\item{ALA.download.data}{(dataframe) A dataframe holding the downloaded ALA data.}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string), A name to use in saving the outputs. Default: 'filtered_data'.}

\item{domain.mask, }{(raster layer) A raster layer specifying the analysis domain}

\item{earliest.year}{(integer) The earliest year for which data will be retained  Default: 1970.}

\item{spatial.uncertainty.m}{(float) The distance (m) for spatial uncertainty below which data will be retained. Default: 2000.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\value{
Dataframe.
}
\description{
Filters data downloaded from the ALA, producing a refined dataset.
}
\examples{
output = filter_ALA_data(My.ALA.data, output.folder = 'C:/Users/processed_data', output.name = My.filtered.ALA.data, domain.mask = Aust.ras)

}
