% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm_builder.R
\name{gdm_builder}
\alias{gdm_builder}
\title{GDM builder}
\usage{
gdm_builder(site.env.data, composition.data, geo = TRUE,
  train.proportion = 0.8, n.pairs.train = NULL, n.pairs.test = NULL,
  n.crossvalid.tests = 10, correlation.threshold = 0.7,
  selection.metric = "D2", sample.method = "random",
  Indiv.Dev.Explained.Min = 1, n.predictors.min = 8, b.used.factor = 2,
  b.dpair.factor = 0.5, b.epair.factor = 1, sigma.spair = NULL,
  spair.factor = 1, domain.mask = NULL, pcs.projargs = NULL,
  bandwidth.geowt = NULL, output.folder = NULL,
  output.name = "gdm_builder_output", verbose = TRUE, ...)
}
\arguments{
\item{site.env.data}{(dataframe) A dataframe holding the location and environmental conditions for each site (grid cell) for which we have species composition data.}

\item{composition.data}{(dataframe) A dataframe holding the final species composition data: the records of all species in all grid cells to be used for modelling.}

\item{geo}{(boolean) Whether geographic distance should be considered as a predictor when deriving the GDM. (default = TRUE)}

\item{train.proportion}{(float) The proportion of sites in 'site.env.data' to use in training the GDM, with the remaining proportion used to test the model. (default = 0.8)}

\item{n.pairs.train}{(integer) The number of site-pairs to use in training the GDM. If not specified, the default is to use 10 percent of the total number of pairs possible from the training sites. (default = NULL)}

\item{n.pairs.test}{(integer) The number of site-pairs to use in testing the GDM. If not specified, the default is to use the same ratio of site-pairs to availabel sites as was used to train the model. (default = NULL)}

\item{n.crossvalid.tests}{(integer) The number of cross-validation sets to use in deriving the GDM. (default = 10)}

\item{correlation.threshold}{(float) The maximum correlation (Pearson's R) permitted between candidate predictor variables, as derived from the sites in 'site.env.data'. (default = 0.7)}

\item{selection.metric}{(string) The model test metric to use in backward elimination of variables for model selection. Options are 'D2' (deviance explained), 'RMSE' (root mean square error), or 'equalised RMSE'. (default = 'D2')}

\item{sample.method}{(string) The site-pair sample method to use. Options are 'random', 'geodist' (geographic distance), 'envdist' (environmental distance), 'geodens' (geographic density), 'geowt' (geographically weighted). (default = 'random')}

\item{Indiv.Dev.Explained.Min}{(float) The minimum amount of deviance explained when potential predictors are assessed individually. (default = 1.0)}

\item{n.predictors.min}{(integer) The target (or minimum) number of predictor variables in the final model. (default = 8)}

\item{b.used.factor}{(float) Multiplier for the decay curve parameter (x-value at curve inflection point) for the number of times each site is used in selected pairs. This factor is multiplied by the ratio of n.pairs.target:n.sites in site.env.data to obtain the b.used parameter (default = 2)}

\item{b.dpair.factor}{(float) For sample method 'geodist'. Multiplier for the decay curve parameter (x-value at curve inflection point) for the distance between sites in a pair. This factor is multiplied by the mean pairwise distance to obtain the b.dpairparameter. (default = 0.5)}

\item{b.epair.factor}{(float) For sample method 'envdist'. Multiplier for the decay curve parameter (x-value at curve inflection point) for the distance between sites in a pair. This factor is multiplied by the mean environmental distance to obtain the b.epair parameter (default = 1)}

\item{sigma.spair}{(float) The standard deviation of the isotropic smoothing kernel used in the 'density()' function from the spatstat package, which is used to determine the density of other sites around each site with compositional data. (default = NULL, in which case the value is set to 5% of the total x-axis extent of 'domain.mask')}

\item{spair.factor}{(float) For sample method 'geodens'. Multiplier for the decay curve parameter (x-value at curve inflection point) for the density of other sites around each site. This factor is multiplied by the mean density of sites to obtain the b.spair parameter. (default = 1.0)}

\item{domain.mask}{(raster layer) A raster layer specifying the analysis domain}

\item{pcs.projargs}{(character) A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation. Used to undertake spatial distance calculations, for example when 'domain.mask' is in geographic coordinate system. For Australian Albers, pcs.projargs="+init=epsg:3577". (default = NULL, in which case the CRS of 'domain.mask' is used for distance calculations).}

\item{bandwidth.geowt}{(float) The bandwidth to use in the 'geowt' (geographically weighted) sample function. (default = NULL, in which case bandwidth is 5% of the x-axis extent)}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. (default = 'gdm_builder_output')}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\value{
List, including GDM model object and cross validation stats.
}
\description{
Derive a GDM for a dataset, including automated variable selection based on cross-validation predictive performance. Produces a final GDM object (average parameters across site-pair samples) and associated cross-validation test metrics.
}
\examples{
output = gdm_builder(My.site.env.data, My.composition.data, output.folder = 'C:/Users/processed_data', output.name = 'My.sitepair.data')

}
