% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm_SingleCrossValidation.R
\name{gdm_SingleCrossValidation}
\alias{gdm_SingleCrossValidation}
\title{Single GDM cross-validation test}
\usage{
gdm_SingleCrossValidation(GDM_Table_Training_In, GDM_Table_Testing_In,
  geo = FALSE, splines = NULL, knots = NULL)
}
\arguments{
\item{GDM_Table_Training_In}{(dataframe) A dataframe holding the GDM input table for model fitting.}

\item{GDM_Table_Testing_In}{(dataframe) A dataframe holding the GDM input table for model testing, having identical column names to 'GDM_Table_Training_In' but using different site-pairs.}

\item{geo}{(boolean) Geographic distance to be used in model fitting (default = FALSE) .}

\item{splines}{(vector) An optional vector of the number of I-spline basis functions to be used for each predictor in fitting the model.}

\item{knots}{(vector) An optional vector of knots in units of the predictor variables to be used in the fitting process.}
}
\description{
Undertake a cross-validation assessment of a GDM, using a single training and testing dataset.
}
\examples{
output = gdm_SingleCrossValidation(My.training.table, My.testing.table, geo=TRUE)

}
