% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdmEngine.R
\name{gdmEngine}
\alias{gdmEngine}
\title{Package containing functions for working with GDM}
\usage{
gdmEngine()
}
\description{
Package containing functions for working with GDM.
}
\section{Functions}{

\code{\link{download_occurrences}}
\code{\link{download_taxalist}}
\code{\link{extract_env_data}}
}

\examples{
?gdmEngine gdmEngine()

}
