% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitr.R
\name{gitr.config}
\alias{gitr.config}
\title{Configure git call}
\usage{
gitr.config(...)
}
\arguments{
\item{...}{List of named args to pass.}
}
\value{
list of git config options
}
\description{
Configure parameters to push file to git, or save parameters for future. Defaults set up for gdm project
}
\note{
make general...?
}
\section{Valid options}{

\itemize{
 \item{"shell_exe"}
 \item{"repo_root"}
 \item{"repo"}
 \item{"files"}
 \item{"msg"}
 \item{"verbose"}
}
}

\examples{
gitr.config()

}
