% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitr.R
\name{gitr.push}
\alias{gitr.push}
\title{Push files to git}
\usage{
gitr.push(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{Print standard output from git. Default TRUE}

\item{repo}{File path to any repository. Default is set to the gdm_workflow repo. Any other path will currently error.}

\item{files}{Files to be tracked. Default: this file (GDM_Workflow_functions.R). 
Paths to a file should be given, and should be relative to the repository root. Also accepts an argument 'all'.}

\item{msg}{Commit message (string) -m flag is set for this git function, so an argument is required. 
Defaults are set provided if left blank.}
}
\value{
std.output from git
}
\description{
Wrappers around git functions. Defaults set up for gdm project
}
\note{
make general...?
}
\examples{
gitr.push()
gitr.push(files = 'all')

}
