% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_downloads.R
\name{merge_downloads}
\alias{merge_downloads}
\title{Merge datasets downloaded from the ALA}
\usage{
merge_downloads(src, keep_unzip = FALSE, parallel = TRUE, verbose = TRUE)
}
\arguments{
\item{src}{(string) Filepath to directory where loadeds are saved. Must not be other .zip files stored in src.}

\item{keep_unzip}{(boolean) Save the unzipped contents of each file?}

\item{parallel}{(boolean) If TRUE (default) will use all CPU (-2) available to load data.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\description{
Unzip and load CSV files downloaded from the ALA.
}
\examples{
output = merge_downloads('C:/Users/raw_files')

}
