% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnls_fit.R
\name{nnls_fit}
\alias{nnls_fit}
\title{Non-negative least squares fitting.}
\usage{
nnls_fit(x, y, weights = rep(1, nobs), start = NULL, etastart = NULL,
  mustart = NULL, offset = rep(0, nobs), family = gaussian(),
  control = list(), intercept = TRUE)
}
\arguments{
\item{x}{}

\item{y}{}

\item{weights}{Default: rep(1, nobs)}

\item{start}{Default: NULL}

\item{etastart}{Default: NULL}

\item{mustart}{Default: NULL}

\item{offset}{Default: rep(0, nobs)}

\item{family}{Default: gaussian()}

\item{control}{Default: list()}

\item{intercept}{Default: TRUE}
}
\value{
list
}
\description{
Function to apply non-negative least squares fitting for a GLM. Code from Andrew Hoskins (CSIRO).
}
\examples{
...
}
