% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outersect}
\alias{outersect}
\title{Get the non-matching objects from two vectors}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{(vector)}

\item{y}{(vector)}
}
\value{
vector
}
\description{
Returns objects not common among two vectors. The opposite to intersect.
}
\note{
https://www.r-bloggers.com/outersect-the-opposite-of-rs-intersect-function/
}
\examples{
outersect(LETTERS[1:10], LETTERS[5:20])
}
