% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PairsDissim}
\alias{PairsDissim}
\title{Compositional dissimilarity calculation for specified pairs of sites}
\usage{
PairsDissim(site_spp, pair_rows, site_rich, max_richness)
}
\arguments{
\item{site_spp}{(matrix) A matrix of the indices for species (col2) occurring in each site (col1), with each row an occurrence}

\item{pair_rows}{(matrix) A matrix of the site index for site 1 (col 1) and site 2 (col 2) in each pair of sites (rows)}

\item{site_rich}{(vector) A vector giving the total number of species in each site (element)}

\item{max_richness}{(integer)}
}
\value{
Vector, the Sorensen dissimilarity between the specified pairs.
}
\description{
Calculate the dissimilarity between specified pairs of sites.
}
\examples{
output = PairsDissim(site.spp, ij.pairs, site.rich, max.rich)
}
