% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PairsDist}
\alias{PairsDist}
\title{Distance calculation for specified pairs}
\usage{
PairsDist(env_vars, pair_rows)
}
\arguments{
\item{env_vars}{(matrix) A matrix of the standardised environment variables (cols) for each site (rows)}

\item{pair_rows}{(matrix) A matrix of the row index in 'env_vars' for site 1 (col 1) and site 2 (col 2) in each pair of sites (rows)}
}
\value{
Vector, the Euclidean distance between the specified pairs.
}
\description{
Calculate the distance between specified pairs.
}
\examples{
output = PairsDist(std.site.env.vars, ij.pairs)
}
