% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pts.euc.distance}
\alias{pts.euc.distance}
\title{Distance between two points}
\usage{
pts.euc.distance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{(float) The x-coordinate of the first point}

\item{y1}{(float) The y-coordinate of the first point}

\item{x2}{(float) The x-coordinate of the second point}

\item{y2}{(float) The y-coordinate of the second point}
}
\value{
(float) The distance between x1,y1 and x2,y2
}
\description{
Calculates the euclidean distance between two points
}
\examples{
my.dist = pts.euc.distance(x1=0, y1=0, x2=3, y2=3)
}
