% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_gridcells_composition.R
\name{select_gridcells_composition}
\alias{select_gridcells_composition}
\title{Select grid cells composition}
\usage{
select_gridcells_composition(ALA.aggregated.data, domain.mask,
  min.richness.threshold = NULL, max.richness.threshold = NULL,
  reference.radius.ncells = NULL, min.proportion.max.richness = 0.25,
  output.folder = NULL, output.name = "selected_gridcell_composition",
  verbose = TRUE)
}
\arguments{
\item{ALA.aggregated.data}{(dataframe) A dataframe of appropriate format held in R working environment.}

\item{domain.mask}{(raster layer) A raster layer specifying the analysis domain}

\item{min.richness.threshold}{(integer) A minimum threshold number of species for selected grid cells. Default: NULL.}

\item{max.richness.threshold}{(integer) A maximum threshold number of species for selected grid cells.  Default: NULL.}

\item{reference.radius.ncells}{(float) A radius (in grid cells) to use in assessing relative richness to maximum.  Default: NULL.}

\item{min.proportion.max.richness}{(float) The minimum threshold proportion of max richness in reference area. Default: 0.25.}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string), A name to use in saving the outputs. Default: 'selected_gridcell_composition'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}

\item{agg.radius.ncells}{(float) A spatial radius, specified as number of grid cells, over which to aggregate records. If NULL, no radial aggregation is performed. Default: NULL.}
}
\value{
Dataframe. Also writes this dataframe to file if output.folder is specified.
}
\description{
Select grid cells with sufficient records for modelling composition.
}
\examples{
output = aggregate_data(my.data, my.raster)

}
