% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepair_sample_density.R
\name{sitepair_sample_density}
\alias{sitepair_sample_density}
\title{Sample-density based site-pair sampler}
\usage{
sitepair_sample_density(site.env.data, n.pairs.target, n.pairs.sample = NULL,
  domain.mask, a.used = 0.05, b.used.factor = 2, c.used = 3,
  sigma.spair = 0.5, a.spair = 0.05, b.spair.factor = 1, c.spair = 1,
  output.folder = NULL, output.name = "site_pairs_data_dens",
  verbose = FALSE)
}
\arguments{
\item{site.env.data}{(dataframe) A dataframe holding the location and environmental conditions for each site (grid cell) for which we have species composition data.}

\item{n.pairs.target}{(integer) The number of site-pairs to select.}

\item{n.pairs.sample}{(integer) The number of pairs to assess simultaneously as part of the sampling process}

\item{domain.mask}{(raster layer) A raster layer specifying the analysis domain}

\item{a.used}{(float) The decay curve parameter (minimum y-value) for the number of times each site is used in selected pairs (default = 0.05)}

\item{b.used.factor}{(float) Multiplier for the decay curve parameter (x-value at curve inflection point) for the number of times each site is used in selected pairs. This factor is multiplied by the ratio of n.pairs.target:n.sites in site.env.data to obtain the b.used parameter (default = 2)}

\item{c.used}{(float) The decay curve parameter (slope of the curve) for the number of times each site is used in selected pairs. (default = 3)}

\item{sigma.spair}{(float) The standard deviation of the isotropic smoothing kernel used in the 'density()' function from the spatstat package, which is used to determine the density of other sites around each site with compositional data. (default = 0.5)}

\item{a.spair}{(float) The decay curve parameter (minimum y-value) for the number of sites within the neighbourhood radius (default = 0.05)}

\item{b.spair.factor}{(float) Multiplier for the decay curve parameter (x-value at curve inflection point) for the density of other sites around each site. This factor is multiplied by the mean density of sites to obtain the b.spair parameter. (default = 1.0}

\item{c.spair}{(float) The decay curve parameter (slope of the curve) for the number of sites within the neighbourhood radius (default = 3)}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. Default: 'site_pairs_data_dens'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\description{
Sample site-pairs based on the number of nearby sites with composition data.
}
\examples{
output = sitepair_sample_density(My.site.env.data, n.pairs.target=10000, domain.mask=My.mask)

}
