% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepair_sample_random.R
\name{sitepair_sample_random}
\alias{sitepair_sample_random}
\title{Random site-pair sampler}
\usage{
sitepair_sample_random(site.env.data, n.pairs.target, output.folder = NULL,
  output.name = "site_pairs_data", verbose = FALSE)
}
\arguments{
\item{site.env.data}{(dataframe) A dataframe holding the location and environmental conditions for each site (grid cell) for which we have species composition data.}

\item{n.pairs.target}{(integer) The number of site-pairs to select.}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. Default: 'pairs_table_dissim'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\description{
Sample site-pairs randomly.
}
\examples{
output = sitepair_sample_random(My.site.env.data, n.pairs.target=10000, output.folder = 'C:/Users/processed_data', output.name = 'My.sitepair.data')

}
