% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzipper}
\alias{unzipper}
\title{Unzip file downloaded from the ALA and read CSV into memory}
\usage{
unzipper(fn, memory_only = TRUE)
}
\arguments{
\item{fn}{(string) Filepath to .zip archive.}

\item{memory_only}{(boolean) If TRUE (default) will delete unzipped data once read into memory}
}
\value{
data.frame
}
\description{
Unzips file and handles exceptions.
}
\examples{
output = unzipper('C:/Users/species_x.zip')
}
