% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm_builder.R
\name{gdm_builder}
\alias{gdm_builder}
\title{GDM builder}
\usage{
gdm_builder(site.env.data, composition.data, geo = TRUE,
  train.proportion = 0.8, n.pairs.train = NULL, n.pairs.test = NULL,
  n.crossvalid.tests = 10, correlation.threshold = 0.7,
  selection.metric = "RMSE", Indiv.Dev.Explained.Min = 1,
  n.predictors.min = 8, output.folder = NULL,
  output.name = "gdm_builder_output", verbose = TRUE)
}
\arguments{
\item{site.env.data}{(dataframe) A dataframe holding the location and environmental conditions for each site (grid cell) for which we have species composition data.}

\item{composition.data}{(dataframe) A dataframe holding the final species composition data: the records of all species in all grid cells to be used for modelling.}

\item{geo}{(boolean) Whether geographic distance should be considered as a predictor when deriving the GDM. (default = TRUE)}

\item{train.proportion}{(float) The proportion of sites in 'site.env.data' to use in training the GDM, with the remaining proportion used to test the model. (default = 0.8)}

\item{n.pairs.train}{(integer) The number of site-pairs to use in training the GDM. If not specified, the default is to use 10 percent of the total number of pairs possible from the training sites. (default = NULL)}

\item{n.pairs.test}{(integer) The number of site-pairs to use in testing the GDM. If not specified, the default is to use the same ratio of site-pairs to availabel sites as was used to train the model. (default = NULL)}

\item{n.crossvalid.tests}{(integer) The number of cross-validation sets to use in deriving the GDM. (default = 10)}

\item{correlation.threshold}{(float) The maximum correlation (Pearson's R) permitted between candidate predictor variables, as derived from the sites in 'site.env.data'. (default = 0.7)}

\item{selection.metric}{(float) The model test metric to use in backward elimination of variables for model selection. Options are 'RMSE' or 'equalised RMSE'. (default = 'RMSE')}

\item{Indiv.Dev.Explained.Min}{(float) The minimum amount of deviance explained when potential predictors are assessed individually. (default = 1.0)}

\item{n.predictors.min}{(integer) The target (or minimum) number of predictor variables in the final model. (default = 8)}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. (default = 'gdm_builder_output')}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\value{
List, including GDM model object and cross validation stats.
}
\description{
Derive a GDM for a dataset, including automated variable selection based on cross-validation predictive performance. Produces a final GDM object (average parameters across site-pair samples) and associated cross-validation test metrics.
}
\examples{
output = gdm_builder(My.site.env.data, My.composition.data, output.folder = 'C:/Users/processed_data', output.name = 'My.sitepair.data')

}
