% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepair_sample_environment.R
\name{sitepair_sample_environment}
\alias{sitepair_sample_environment}
\title{Environment based site-pair sampler}
\usage{
sitepair_sample_environment(site.env.data, n.pairs.target, env.colnames,
  dist.method = "euclidean", n.pairs.sample = NULL, a.used = 0.05,
  b.used.factor = 2, c.used = 3, a.epair = 0.05, b.epair.factor = 1,
  c.epair = 3, output.folder = NULL, output.name = "site_pairs_data_env",
  verbose = FALSE)
}
\arguments{
\item{site.env.data}{(dataframe) A dataframe holding the location and environmental conditions for each site (grid cell) for which we have species composition data.}

\item{n.pairs.target}{(integer) The number of site-pairs to select.}

\item{env.colnames}{(character vector) A list of the column names in 'site.env.data' to use in determining environmental distance.}

\item{dist.method}{(character) Not used yet - only Euclidean distance applied}

\item{n.pairs.sample}{(integer) The number of pairs to assess simultaneously as part of the sampling process}

\item{a.used}{(float) The decay curve parameter (minimum y-value) for the number of times each site is used in selected pairs (default = 0.05)}

\item{b.used.factor}{(float) Multiplier for the decay curve parameter (x-value at curve inflection point) for the number of times each site is used in selected pairs. This factor is multiplied by the ratio of n.pairs.target:n.sites in site.env.data to obtain the b.used parameter (default = 2)}

\item{c.used}{(float) The decay curve parameter (slope of the curve) for the number of times each site is used in selected pairs. (default = 3)}

\item{a.epair}{(float) The decay curve parameter (minimum y-value) for the distance between sites in a pair (default = 0.05)}

\item{b.epair.factor}{(float) Multiplier for the decay curve parameter (x-value at curve inflection point) for the distance between sites in a pair. This factor is multiplied by the mean environmental distance to obtain the b.epair parameter (default = 1)}

\item{c.epair}{(float) The decay curve parameter (slope of the curve) for the distance between sites in a pair.  (default = 3)}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. Default: 'site_pairs_data_geo'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\description{
Sample site-pairs based on the environmental distance between sites.
}
\examples{
output = sitepair_sample_environment(My.site.env.data, n.pairs.target=10000)

}
