% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepair_sample_geo_weighted.R
\name{sitepair_sample_geo_weighted}
\alias{sitepair_sample_geo_weighted}
\title{Geographically weighted site-pair sampler}
\usage{
sitepair_sample_geo_weighted(site.env.data, n.pairs.target, bandwidth = NULL,
  domain.mask, pcs.projargs = NULL, output.folder = NULL,
  output.name = "site_pairs_data_geowt", verbose = FALSE)
}
\arguments{
\item{site.env.data}{(dataframe) A dataframe holding the location and environmental conditions for each site (grid cell) for which we have species composition data.}

\item{n.pairs.target}{(integer) The number of site-pairs to select.}

\item{bandwidth}{(float) The bandwidth to use in the 'geowt' (geographically weighted) sample function. (default = NULL, in which case bandwidth is 5% of the x-axis extent)}

\item{domain.mask}{(raster layer) A raster layer specifying the analysis domain}

\item{pcs.projargs}{(character) A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation. Used to undertake spatial distance calculations, for example when 'domain.mask' is in geographic coordinate system. (default = NULL, in which case the CRS of 'domain.mask' is used for distance calculations).}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. Default: 'site_pairs_data_dens'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\value{
Dataframe, site-pairs table, being first 6 columns of the GDM input table, with dissimilarities not calculated.
}
\description{
Sample site-pairs based on the number of nearby sites with composition data.
}
\examples{
output = sitepair_sample_density(My.site.env.data, n.pairs.target=10000, domain.mask=My.mask)

}
