% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitepair_sample_geographic.R
\name{sitepair_sample_geographic}
\alias{sitepair_sample_geographic}
\title{Geographic based site-pair sampler}
\usage{
sitepair_sample_geographic(site.env.data, n.pairs.target,
  n.pairs.sample = NULL, a.used = 0.05, b.used = NULL, c.used = 3,
  a.dpair = 0.05, b.dpair = NULL, c.dpair = 3, output.folder = NULL,
  output.name = "site_pairs_data_geo", verbose = FALSE)
}
\arguments{
\item{site.env.data}{(dataframe) A dataframe holding the location and environmental conditions for each site (grid cell) for which we have species composition data.}

\item{n.pairs.target}{(integer) The number of site-pairs to select.}

\item{n.pairs.sample}{(integer) The number of pairs to assess simultaneously as part of the sampling process}

\item{a.used}{(float) The decay curve parameter (minimum y-value) for the number of times each site is used in selected pairs (default = 0.05)}

\item{b.used}{(float) The decay curve parameter (x-value at curve inflection point) for the number of times each site is used in selected pairs (default = 2 x the total number of sites in 'site.env.data')}

\item{c.used}{(float) The decay curve parameter (slope of the curve) for the number of times each site is used in selected pairs (default = 3)}

\item{a.dpair}{(float) The decay curve parameter (minimum y-value) for the distance between sites in a pair (default = 0.05)}

\item{b.dpair}{(float) The decay curve parameter (x-value at curve inflection point) for the distance between sites in a pair (default = half the maximum distance between sites)}

\item{c.dpair}{(float) The decay curve parameter (slope of the curve) for the distance between sites in a pair (default = 3)}

\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. Default: 'site_pairs_data_geo'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}
}
\description{
Sample site-pairs based on the geographic distance between sites.
}
\examples{
output = sitepair_sample_geographic(My.site.env.data, n.pairs.target=10000)

}
