% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformGrids.R
\name{TransformGrids}
\alias{TransformGrids}
\title{Transform grids}
\usage{
TransformGrids(gdm.model, env.grids.stk, output.folder = NULL,
  output.name = "tran", verbose = TRUE)
}
\arguments{
\item{output.folder}{(string) A folder to save the outputs to. If none specified, no file is written.}

\item{output.name}{(string) A name to use in saving the outputs. Default: 'pairs_table_dissim'.}

\item{verbose}{(boolean) Print messages to console. Default TRUE.}

\item{pairs.table}{(dataframe) A dataframe holding the site-pairs. This is the first six columns of a gdm input table, plus 4 columns hold s1 & s2 long & lat.}

\item{composition.data}{(dataframe) A dataframe holding the final species composition data: the records of all species in all grid cells to be used for modelling.}
}
\value{
Dataframe.
}
\description{
The R function to call in order to transform big grids. This in turn calls a Rcpp function, that does the magic
}
\examples{
output = calculate_dissimilarities(My.pairs.table, My.composition.data, output.folder = 'C:/Users/processed_data', output.name = 'My.sitepair.data')

}
